Move TizenDeviceManager into a separate file
